separate some tests
